batcheval: move ExportRequest to batcheval
Release justificaiton: very low risk, pure code move for ease of future backports.
Release note: none.
