+++ pkg/ccl/backupccl/BUILD.bazel
@@ -48,6 +48,7 @@ go_library(
         "//pkg/kv",
         "//pkg/kv/bulk",
         "//pkg/kv/kvserver",
+        "//pkg/kv/kvserver/batcheval",
         "//pkg/kv/kvserver/concurrency/lock",
         "//pkg/kv/kvserver/protectedts",
         "//pkg/roachpb:with-mocks",
