+++ pkg/ccl/storageccl/BUILD.bazel
@@ -4,7 +4,6 @@ go_library(
     name = "storageccl",
     srcs = [
         "encryption.go",
-        "export.go",
         "external_sst_reader.go",
         "import.go",
         "revision_reader.go",
@@ -14,12 +13,8 @@ go_library(
     deps = [
         "//pkg/base",
         "//pkg/cloud",
-        "//pkg/keys",
         "//pkg/kv",
         "//pkg/kv/kvserver",
-        "//pkg/kv/kvserver/batcheval",
-        "//pkg/kv/kvserver/batcheval/result",
-        "//pkg/kv/kvserver/spanset",
         "//pkg/roachpb:with-mocks",
         "//pkg/settings",
         "//pkg/settings/cluster",
@@ -27,11 +22,9 @@ go_library(
         "//pkg/util/hlc",
         "//pkg/util/humanizeutil",
         "//pkg/util/retry",
-        "//pkg/util/tracing",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_pebble//sstable",
         "@com_github_cockroachdb_pebble//vfs",
-        "@com_github_gogo_protobuf//types",
         "@org_golang_x_crypto//pbkdf2",
     ],
 )
@@ -41,31 +34,19 @@ go_test(
     size = "large",
     srcs = [
         "encryption_test.go",
-        "export_test.go",
         "external_sst_reader_test.go",
         "main_test.go",
     ],
     embed = [":storageccl"],
     deps = [
-        "//pkg/base",
-        "//pkg/cloud/impl:cloudimpl",
-        "//pkg/keys",
-        "//pkg/kv",
-        "//pkg/roachpb:with-mocks",
         "//pkg/security",
         "//pkg/security/securitytest",
         "//pkg/server",
-        "//pkg/settings/cluster",
-        "//pkg/storage",
-        "//pkg/testutils",
         "//pkg/testutils/serverutils",
-        "//pkg/testutils/sqlutils",
         "//pkg/testutils/testcluster",
-        "//pkg/util/hlc",
         "//pkg/util/humanizeutil",
         "//pkg/util/leaktest",
         "//pkg/util/randutil",
-        "@com_github_cockroachdb_errors//:errors",
         "@com_github_stretchr_testify//require",
     ],
 )
