+++ pkg/kv/kvserver/batcheval/BUILD.bazel
@@ -11,6 +11,7 @@ go_library(
         "cmd_delete.go",
         "cmd_delete_range.go",
         "cmd_end_transaction.go",
+        "cmd_export.go",
         "cmd_gc.go",
         "cmd_get.go",
         "cmd_heartbeat_txn.go",
@@ -97,6 +98,7 @@ go_test(
         "cmd_add_sstable_test.go",
         "cmd_clear_range_test.go",
         "cmd_end_transaction_test.go",
+        "cmd_export_test.go",
         "cmd_get_test.go",
         "cmd_lease_test.go",
         "cmd_query_resolved_timestamp_test.go",
@@ -132,6 +134,7 @@ go_test(
         "//pkg/testutils",
         "//pkg/testutils/serverutils",
         "//pkg/testutils/skip",
+        "//pkg/testutils/sqlutils",
         "//pkg/testutils/testcluster",
         "//pkg/util/hlc",
         "//pkg/util/leaktest",
